Make use for 'size --format=GNU' where possible
The size program by default counts read only data as text.  There is a
fix proposed for this, but it is not merged, and even if/when it is
merged most users will not have the new version immediately.

This change probes for the availability of the 'size --format=GNU'
feature (which is the new flag to count read only data as data), if
the flag is available use this, otherwise, use the default format and
accept a slight inaccuracy in the results.

ChangeLog:

	* testsuite/lib/beebs.exp (beebs_size_all_benchmarks): Use 'size
	--format=GNU' where possible.