Fix testsuite.
This fixes two issues:

- An error in the testsuite for the "none" board - required
functions did not exist. Now the required functions exist, and all
execution tests are marked as unsupported.
- Non-existent tests (e.g. matmult) were getting marked as unresolved.

This required the removal of the beebs_restore_default_procs
function, which makes it impossible to set beebs_* functions on a
board-by-board basis, and does not appear to have a useful
purpose.

testsuite/ChangeLog:

        * config/none.exp (non_filter_benchmark, none_init): New functions.
        * lib/beebs.exp: Rename default_beebs_* to beebs_*.
        (beebs_run_benchmark): Make it OK not to find an executable.
        (beebs_restore_default_procs): Delete.
        (beebs_init): Don't call beebs_restore_default_procs.